nrl-scraper-final/requirements.txt
# NRL Scraper - Runtime Dependencies (SPEC-1)
# ============================================

# HTTP & Retries
httpx>=0.27,<0.29
tenacity>=8.2,<9

# Validation
pydantic>=2.6,<3

# Data Processing
pandas>=2.2,<3
pyarrow>=15,<20

# Database (PostgreSQL 15 via psycopg3)
sqlalchemy>=2.0,<2.1
psycopg[binary,pool]>=3.2,<4

# HTML Parsing
beautifulsoup4>=4.12,<5
lxml>=5.2,<6

# Date Handling
python-dateutil>=2.8,<3

# Environment
python-dotenv>=1.0,<2
